requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
pyyaml>=6.0
tqdm>=4.62.0
scikit-learn>=1.0.0  # For decision tree example
joblib>=1.1.0  # For parallel processing
